Add Apache 2.0 License
This was a choice based on that pph-theme is public (I think it has to
be unless we can provision GitHub access to production servers) and the
fact that it includes limitations on use of trademark(s).

I added this as a separate commit so we can easily drop or amend it.